Recreated Solution to fix CI
