--- csharp-advanced-topics/ActionAndFuncInCSharp/ActionAndFunc.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionAndFunc", "ActionAndFunc\ActionAndFunc.csproj", "{896EB0D5-E0AC-49AD-9A4D-81BAA2FEE8D3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionAndFunc.Tests", "ActionAndFunc.Tests\ActionAndFunc.Tests.csproj", "{072EAFCF-144A-490B-A151-F45384CB9738}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{896EB0D5-E0AC-49AD-9A4D-81BAA2FEE8D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{896EB0D5-E0AC-49AD-9A4D-81BAA2FEE8D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{896EB0D5-E0AC-49AD-9A4D-81BAA2FEE8D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{896EB0D5-E0AC-49AD-9A4D-81BAA2FEE8D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{072EAFCF-144A-490B-A151-F45384CB9738}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{072EAFCF-144A-490B-A151-F45384CB9738}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{072EAFCF-144A-490B-A151-F45384CB9738}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{072EAFCF-144A-490B-A151-F45384CB9738}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
